requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
faiss-cpu==1.13.2
openai==1.3.0
pypdf==3.17.1
python-dotenv==1.0.0
streamlit==1.28.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
numpy>=2.0.0
